Hacky build fixes

+++ OpenHome/Exception.cpp
@@ -41,7 +41,7 @@ static void CallFatalErrorHandler(const char* aMsg)
 void OpenHome::AssertHandlerDefault(const TChar* aFile, TUint aLine)
 {
     char buf[1024];
-    (void)snprintf(buf, sizeof(buf), "Assertion failed.  %s:%lu\n", aFile, aLine);
+    (void)snprintf(buf, sizeof(buf), "Assertion failed.  %s:%lu\n", aFile, (unsigned long)aLine);
     CallFatalErrorHandler(buf);
     Os::Quit();
 }
@@ -67,7 +67,7 @@ void OpenHome::UnhandledExceptionHandler(const TChar* aExceptionMessage, const T
     GetThreadName(thName);
     char buf[1024];
     Log::Print("boo\n");
-    (void)snprintf(buf, sizeof(buf), "Unhandled exception %s at %s:%lu in thread %s\n", aExceptionMessage, aFile, aLine, thName.Ptr());
+    (void)snprintf(buf, sizeof(buf), "Unhandled exception %s at %s:%lu in thread %s\n", aExceptionMessage, aFile, (unsigned long)aLine, thName.Ptr());
     CallFatalErrorHandler(buf);
 }
 
@@ -78,7 +78,7 @@ void OpenHome::UnhandledExceptionHandler(Exception& aException)
     char buf[1024];
     Log::Print("woo\n");
     (void)snprintf(buf, sizeof(buf), "Unhandled exception %s at %s:%lu in thread %s\n",
-                   aException.Message(), aException.File(), aException.Line(), thName.Ptr());
+                   aException.Message(), aException.File(), (unsigned long)aException.Line(), thName.Ptr());
     CallFatalErrorHandler(buf);
 }
