+++ OpenHome/Net/Device/Upnp/DviServerWebSocket.cpp
@@ -975,7 +975,7 @@ SocketTcpServer* DviServerWebSocket::CreateServer(const NetworkAdapter& aNif)
     TChar thName[5];
 	const TUint numWsThreads = Stack::InitParams().DvNumWebSocketThreads();
     for (TUint i=0; i<numWsThreads; i++) {
-        (void)sprintf(&thName[0], "WS%2lu", i);
+        (void)sprintf(&thName[0], "WS%2lu", (unsigned long)i);
         server->Add(&thName[0], new DviSessionWebSocket(aNif.Address(), server->Port()));
     }
     return server;
